improve pagerank performance
- stop pagerank iterations if the results have converged to a chosen
  epsilon

+++ pagerank.hpp
@@ -12,7 +12,7 @@ public:
      * Create a new PageRank object given a list of edges (represented as pairs
      * of node ids). Damping should usually be 0.85
      */
-    PageRank(const vector<pair<size_t, size_t>> edges, int numNodes, double damping);
+    PageRank(const vector<pair<size_t, size_t>> edges, size_t numNodes, double damping);
 
     /**
      * Get the result of the PageRank algorithm. Values are between 0 and 1.
